+++ src/main/java/com/Jguides/spingboot/repository/WarehouseRepository.java
@@ -3,5 +3,5 @@ package com.Jguides.spingboot.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.Jguides.spingboot.Model.Warehouse;
 
-public interface WarehouseRepository extends JpaRepository<Warehouse, Long> {
+public interface WarehouseRepository extends JpaRepository<Warehouse, Integer> {
 }
